move to test_policies
